requirements.txt
tqdm==4.36.1
pymysql==0.10.1
sqlalchemy==1.3.10
mysqlclient==2.0.2
pyexecjs==1.5.1
sentencepiece==0.1.95
matplotlib==3.3.4